# src/webapi/swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Feature:
    description: Feature defines the structure for each feature of a product
    properties:
      code:
        description: the code friendly name of the feature
        type: string
        x-go-name: Code
      id:
        $ref: '#/definitions/ObjectID'
      name:
        description: the user friendly name of the feature
        type: string
        x-go-name: Name
      type:
        $ref: '#/definitions/FlagType'
    required:
    - name
    - code
    - type
    type: object
    x-go-package: github.com/serdarkalayci/goboiler/webapi/data
  FlagType:
    description: FlagType is the enum that enumerates the type of feature flag
    format: int64
    type: integer
    x-go-package: github.com/serdarkalayci/goboiler/webapi/data
  GenericError:
    description: GenericError is a generic error message returned by a server
    properties:
      message:
        type: string
        x-go-name: Message
    type: object
    x-go-package: github.com/serdarkalayci/goboiler/webapi/handlers
  ObjectID:
    items:
      format: uint8
      type: integer
    title: ObjectID is the BSON ObjectID type.
    type: array
    x-go-package: go.mongodb.org/mongo-driver/bson/primitive
  Product:
    description: Product defines the structure for a product
    properties:
      features:
        description: the ISBNFeature list of the product
        items:
          $ref: '#/definitions/Feature'
        type: array
        x-go-name: Features
      id:
        $ref: '#/definitions/ObjectID'
      name:
        description: the name of the product
        type: string
        x-go-name: Name
    required:
    - name
    type: object
    x-go-package: github.com/serdarkalayci/goboiler/webapi/data
...
info:
  description: Documentation for Details API
  title: of Details API
  version: 1.0.0
paths:
...
  /products/{id}:
    get:
      description: Return a list of Product from the database
      operationId: getSingleProduct
      responses:
        "200":
          $ref: '#/responses/ProductResponse'
        "404":
          $ref: '#/responses/errorResponse'
      tags:
      - Products
produces:
- application/json
responses:
...
  ProductResponse:
    description: Data structure representing a single product
    schema:
      $ref: '#/definitions/Product'
...
  errorResponse:
    description: Generic error message returned as a string
    schema:
      $ref: '#/definitions/GenericError'
...
schemes:
- http
swagger: "2.0"